Update benchmarks in readme.txt
Benchmarked with the new more isolated system
using siege. The relative rankings are unchanged,
but throughput is higher with siege than with the
janky parallel hey script.

I left WP Redis out because it had inconsistent
showings. It was sometimes very good and sometimes
very bad. I'm not sure if that was a benchmarking
issue or an issue with WP Redis itself, but I'm
leaving it out for now so I don't misrepresent it.
The other plugins have all been very consistent.